refactor: moved fax configuration tests into FaxConfigurationTests
